split the export tasks
