+++ posthog/tasks/__init__.py
@@ -1,4 +1,6 @@
 # Make tasks ready for celery autoimport
+from posthog.tasks.exports import csv_exporter, insight_exporter
+
 from . import (
     async_migrations,
     calculate_cohort,
@@ -6,7 +8,6 @@ from . import (
     check_clickhouse_schema_drift,
     delete_clickhouse_data,
     email,
-    exporter,
     split_person,
     status_report,
     sync_all_organization_available_features,
@@ -21,7 +22,8 @@ __all__ = [
     "check_clickhouse_schema_drift",
     "delete_clickhouse_data",
     "email",
-    "exporter",
+    "csv_exporter",
+    "insight_exporter",
     "split_person",
     "status_report",
     "sync_all_organization_available_features",
