+++ posthog/tasks/test/test_exporter.py
@@ -3,11 +3,11 @@ from unittest.mock import MagicMock, patch
 
 from posthog.models.dashboard import Dashboard
 from posthog.models.exported_asset import ExportedAsset
-from posthog.tasks.exporter import export_task, get_driver
+from posthog.tasks.exports.insight_exporter import export_insight, get_driver
 from posthog.test.base import APIBaseTest
 
 
-@patch("posthog.tasks.exporter.uuid")
+@patch("posthog.tasks.exports.insight_exporter.uuid")
 class TestExporterTask(APIBaseTest):
     exported_asset: ExportedAsset = None  # type: ignore
 
@@ -24,11 +24,11 @@ class TestExporterTask(APIBaseTest):
         with open("/tmp/posthog_test_exporter.png", "wb") as fh:
             fh.write(base64.decodebytes(example_png))
 
-    @patch("posthog.tasks.exporter.get_driver")
+    @patch("posthog.tasks.exports.insight_exporter.get_driver")
     def test_exporter_runs(self, mock_get_driver: MagicMock, mock_uuid: MagicMock) -> None:
         mock_uuid.uuid4.return_value = "posthog_test_exporter"
         assert self.exported_asset.content is None
-        export_task(self.exported_asset.id)
+        export_insight(self.exported_asset.id)
         self.exported_asset.refresh_from_db()
         assert self.exported_asset.content is not None
